LayoutTests/svg/animations/animate-path-nested-transforms-expected.txt

SVG 1.1 dynamic animation tests

test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getBBox().x is almost 132 (within 20)
PASS rootSVGElement.getBBox().y is almost -90 (within 20)
PASS rootSVGElement.getBBox().x is almost 332 (within 20)
PASS rootSVGElement.getBBox().y is almost 550 (within 20)
PASS successfullyParsed is true

TEST COMPLETE
